### 2022-05-17 #4402 @editorialbot
[PRE REVIEW]: 'HallThruster.jl: a Julia package for 1D Hall thruster discharge simulation'
**Submitting author:** @archermarx (<a href="http://orcid.org/0000-0003-3614-6127">Thomas Marks</a>)
**Repository:** https://github.com/UM-PEPL/HallThruster.jl
**Branch with paper.md** (empty if default branch):
**Version:** v0.1.0
