### 2022-05-18 945f635
fix wall loss bug and add license
