+++ paper/paper.bib
@@ -99,7 +99,7 @@ title = {Fundamentals of Electric Propulsion},
 booktitle = {Fundamentals of Electric Propulsion},
 chapter = {7},
 pages = {325-392},
-doi = {https://doi.org/10.1002/9780470436448.ch7},
+doi = {10.1002/9780470436448.ch7},
 eprint = {https://onlinelibrary.wiley.com/doi/pdf/10.1002/9780470436448.ch7},
 year = {2008},
 keywords = {hall thruster, hall-effect thrustuers (HETs), stationary plasam thrusters (SPTs), hall current, electron temperature},
@@ -131,7 +131,8 @@ eprint = {
   title={High-performance symbolic-numerics via multiple dispatch},
   author={Gowda, Shashi and Ma, Yingbo and Cheli, Alessandro and Gwozdz, Maja and Shah, Viral B and Edelman, Alan and Rackauckas, Christopher},
   journal={arXiv preprint arXiv:2105.03949},
-  year={2021}
+  year={2021},
+  doi={10.1145/3511528.3511535}
 }
 
 @article{bezanson2017julia,
@@ -143,7 +144,8 @@ eprint = {
   pages={65--98},
   year={2017},
   publisher={SIAM},
-  url={https://doi.org/10.1137/141000671}
+  url={https://doi.org/10.1137/141000671},
+  doi={10.1137/141000671},
 }
 
 @article{thomas1949elliptic,
