### 2022-05-18 0cf09d1
fix indexing problem in freq_electron_wall
