### 2022-05-23 f70dfcc
basics of charge exchange and smoothstep transition function
