### 2022-05-25 cc5b1e7
update wall loss documentation
