+++ docs/src/wall_loss_models.md
@@ -4,7 +4,7 @@ HallThruster.jl allows you to choose from three different wall loss models. They
 
 ## Background
 
-The core of the wall loss models in HallThruster.jl is the abstract type `WallLossModel`. It has three children: `NoWallLosses`, `ConstantSheathPotential`, and `WallSheath`. `ConstantSheathPotential` has three fields. A wall `sheath_potential` to be set by the user, and an `inner_loss_coeff` and `outer_loss_coeff` which allow to scale the energy loss inside vs. outside the thruster channel. `WallSheath` has one field `material`, which is of type `WallMaterial`.
+The core of the wall loss models in HallThruster.jl is the abstract type `WallLossModel`. It has three children: `NoWallLosses`, `ConstantSheathPotential`, and `WallSheath`. `ConstantSheathPotential` has three fields. A wall `sheath_potential` to be set by the user, and an `inner_loss_coeff` and `outer_loss_coeff` which allow to scale the energy loss inside vs. outside the thruster channel. `WallSheath` has two field: `material`, which is of type `WallMaterial` and includes information about secondary electron emission yields, and α, which is a constant wall loss scaling coefficient (values around 0.1-0.2 are good usually, but this may need to be calibrated against some data).
 
 ## Provided wall loss models
 
@@ -14,7 +14,51 @@ HallThruster.jl provides three models out of the box. These are
 | ----------------------- | ------------------------------------------------------------ | ------------------------------------------------------------ |
 | `NoWallLosses` | `Any`                                                      | Ignores electron energy losses to the walls. May cause numerical issues.  |
 | `ConstantSheathPotential`         | `Any` | Employs a simple sheath energy loss model with constant sheath potential, based on the electron Boltzmann equation for electron density in the sheath as a function of electron temperature. Uses constants to scale losses inside and outside the thruster. See also [JP Boeuf, *Low frequency oscillations in a stationary plasma thruster*, Journal of Applied Physics 84, 3541, 1998](https://aip.scitation.org/doi/10.1063/1.368529) and [Landmark study](https://www.landmark-plasma.com/test-case-3)|
[…]
