### 2022-06-01 50b2d19
Update physics.md
removed broken links
