+++ CITATION.bib
@@ -1,0 +1,7 @@
+@misc{Marks2022,
+author = {Marks, Thomas A. and Schedler, Paul and Jorns, Benjamin A.},
+doi = {10.5281/zenodo.6614275},
+title = {{HallThruster.jl - An open source 1D fluid Hall thruster code}},
+url = {https://github.com/UM-PEPL/HallThruster.jl},
+year = {2022}
+}
