### 2022-06-22 #4402 @lucydot
Hi @archermarx  - thanks for the suggestions.

@StanczakDominik @akaptano are you available to review this submission for JOSS? I see you both have plasma physics + Julia knowledge 💪
