### 2022-06-27 #4402 @lucydot
@datseris @jameshclrk @matbesancon are any of you available to review this JOSS submission?
