### 2022-07-05 436ce2f
refactor anomalous transport
add fixedanommodel as choice
