### 2022-07-05 0402946
fix bug in Bohm model specification
