### 2022-07-11 9e23517
add documentation for new anom interface
