+++ docs/src/anomalous_transport.md
@@ -1,51 +1,122 @@
 # Anomalous Transport
 
-HallThruster has two anomalous transport models built in and allows users to define their own. Currently, we only support algebraic (zero-equation) models, but support for models involving additional PDEs will be added in the future.
+HallThruster has a few anomalous transport models built in and allows users to define their own.
+Currently, we only support anomalous transport models that are fixed as a function of space or
+algebraic models which depend on the local plasma properties, but multi-equation transport
+models are planned for the future.
 
 ## Built-in Models
 
-### `TwoZoneBohm`
+### `NoAnom()`
+
+Model for no anomalous transport (anomalous collision frequency  = 0).
+
+```julia
+anom_model = NoAnom()
+```
+
+### `Bohm(c)`
+
+Model where the anomalous collision frequency scales with the electron cyclotron frequency ωce times some scaling factor c
+
+### `TwoZoneBohm(c1, c2)`
 
 HallThruster's default anomalous transport option. This is a standard model of anomalous transport frequently used in Hall thruster simulations. The anomalous collision frequency is defined as
 
 ```math
 \nu_{AN} = \begin{cases}
-    \beta_1 \omega_{ce} & z < L_{ch} \\
-    \beta_2 \omega_{ce} & z \ge L_{ch}
+    \c_1 \omega_{ce} & z < L_{ch} \\
+    \c_2 \omega_{ce} & z \ge L_{ch}
 \end{case}
 ```
 
-In the above expression, ``\beta_1`` and ``beta_2`` are tunable coefficients, ``omega_{ce} = e B / m_e`` is the electron cyclotron frequency, and ``L_{ch}`` is the channel length. A `TwoZoneBohm` model is initialized as follows
+In the above expression, ``c_1`` and ``c_2`` are tunable coefficients, ``omega_{ce} = e B / m_e`` is the electron cyclotron frequency, and ``L_{ch}`` is the channel length. A `TwoZoneBohm` model is initialized as follows
 
 ```julia
-anom_model = TwoZoneBohm(β1, β2)
+anom_model = TwoZoneBohm(c1, c2)
 ```
 
-### `NoAnom`
+The transition between the zones is determined by the user-provided transition function.
+This defaults to a step function.
 
-Model for no anomalous transport (anomalous collision frequency  = 0).
+### `MultiLogBohm(z, c)`
+
+Model similar to that employed in Hall2De, where the mobility is Bohm-like (i.e. `νan(z) = c(z) * ωce(z)`) and z is in meters.
+
+The function `c(z)` is defined by a sequence of nodes `(z, c)` provided by the user. At `z = z[1]`, `c(z) = c[1]`, and so forth.
+
+At `z[i] < z < z[i+1]`, `log(c)` is defined by linearly interpolating between `log(c[i])` and `log(c[i+1])`.
+
+For `z < z[1]`, `c = c[1]` and for `z > z[end]`, `c(z) = c[end]`.
+
+The user may also provide a single array of `[z[1], z[2], ..., z[end], c[1], c[2], ..., c[end]]`. The number of `z `values must be equal to the number of c values.
+
+## The `AnomalousTransportModel` interface
+All anomalous transport models in HallThruster.jl are subtypes of `AnomalousTransportModel`.
+Two methods must be defined for all such models.
+
+### `initialize_anom!(νan, model, U, params)`
+This function takes a pre-allocated vector `νan` of size `length(params.z_cell)`, the model,
+the initial simulation state vector `U`, and the simulation `params`, and fills `νan` with
+the anomalous collision frequency at the start of the simulation at each cell center.
+
+### `evaluate_anom(U, params, i)`
+This function takes the state vector `U`, the simulation `params` and cell number `i` and
+returns the anomalous collision frequency for that cell.
+
+## Types of model
+
+### `FixedAnomModel`
[…]
