### 2022-07-11 44c0ebc
remove symbolics as a main dependency
