### 2022-07-13 #4402 @lucydot
@archermarx one option may be to contact some of those you have cited in the JOSS paper: Kentaro Harra, Rupali Sahu or Adnan Mansour. Before I do, can I ask if you have any possible conflict of interest? I know in some fields its hard _not_ to have a conflict of interest with those you are citing!
