### 2022-07-20 87e6ab3
add some convenience functions
