### 2022-07-21 fcfa12f
add new tutorial, efficiency computation
