+++ docs/src/tutorial.md
@@ -1,3 +1,3 @@
-# Tutorial
+# Tutorial notebook
 
 A jupyter notebook tutorial can be found [here](https://nbviewer.org/github/UM-PEPL/HallThruster.jl/blob/main/tutorials/HallThrusterTutorial.ipynb).

[…]
