### 2022-07-21 ac61629
fix bug in wall sheath inner constructor
