### 2022-07-22 7bb7ed7
improve averaging of velocity and density and add to cache
