### 2022-07-22 b24d547
add anode sheath function and remove refs to phi
