### 2022-07-22 2020bcc
overhaul potential and electric field computation (tests not fixed yet)
