### 2022-07-25 f8bbefb
removal of old potential solver and fix tests
