+++ src/simulation/potential.jl
@@ -14,9 +14,7 @@ function anode_sheath_potential(U, params)
     mi = params.config.propellant.m
 
     # Compute anode sheath potential
-    if config.LANDMARK
-        Vs = 0.0
-    else
+    if config.anode_boundary_condition == :sheath
         ce = sqrt(8 * e * params.cache.Tev[1] / π / me)
         je_sheath = e * ne[1] * ce / 4
 
@@ -33,6 +31,8 @@ function anode_sheath_potential(U, params)
         else
             Vs = -params.cache.Tev[1] * log(min(1.0, je_sheath_edge / je_sheath))
         end
+    else
+        Vs = 0.0
     end
 
     return Vs
