### 2022-08-03 8abdc10
first pass background pressure effects
