+++ src/simulation/boundaryconditions.jl
@@ -1,5 +1,5 @@
 function left_boundary_state!(bc_state, U, params)
-    (;index, A_ch, config) = params
+    (;index, A_ch, config,) = params
     mi = config.propellant.m
     (;Tev) = params.cache
 
@@ -28,7 +28,16 @@ function left_boundary_state!(bc_state, U, params)
     bohm_velocity = bohm_factor * sqrt(e * Tev[1] / mi)
 
     # Add inlet neutral density
-    bc_state[index.ρn] = mdot_a / A_ch / un
+    bc_state[index.ρn[1]] = mdot_a / A_ch / un
+
+    if config.solve_background_neutrals
+        # Background neutrals are accomodated and re-emitted as anode neutrals
+        background_neutral_density = U[index.ρn[2], begin+1]
+        background_neutral_velocity = params.background_neutral_velocity
+        background_neutral_flux = -background_neutral_density * background_neutral_velocity
+        bc_state[index.ρn[1]] += background_neutral_flux / un
+        bc_state[index.ρn[2]] = U[index.ρn[2], begin+1]
+    end
 
     @inbounds for Z in 1:params.config.ncharge
         boundary_density = U[index.ρi[Z],   begin+1]
@@ -40,15 +49,23 @@ function left_boundary_state!(bc_state, U, params)
 
         recombination_density = -(boundary_density * boundary_velocity) / un
 
-        bc_state[index.ρn] += recombination_density
+        bc_state[index.ρn[1]] += recombination_density
         bc_state[index.ρi[Z]] = boundary_density # Neumann BC for ion density at left boundary
         bc_state[index.ρiui[Z]] = boundary_velocity * boundary_density
     end
 end
 
 function right_boundary_state!(bc_state, U, params)
-    (;index) = params
-    bc_state[index.ρn] = U[index.ρn, end-1]
+    (;index, fluids) = params
+
+    @inbounds for j in 1:params.num_neutral_fluids
+        # If neutral fluid is right-moving, use Neumann BC, otherwise Dirichlet BC
+        if fluids[j].conservation_laws.u ≥ 0.0
+            bc_state[index.ρn[j]] = U[index.ρn[j], end-1]
+        else
+            bc_state[index.ρn[j]] = params.background_neutral_density
+        end
+    end
 
     @inbounds for Z in 1:params.config.ncharge
         boundary_density = U[index.ρi[Z], end-1]
