### 2022-08-11 c461f3c
json reads bfield in tesla
