### 2022-08-31 #4672 @lucydot
@StanczakDominik - I can see you have started your review - great! - let me know if you have any queries as you progress.

@Rupali-Sahu - your first review task is to generate a checklist - there is more information at the top of this thread. Please, if you have any questions, just ask in the thread below.

I find reviews work best when it is a back-and-forth conversations between reviewers and authors, so please feel free to give your thoughts and suggestions as you are working through the checklist. In many cases, it may best to raise these as an issue on the HallThruster repository - if you do, please let us know in this thread.
