### 2022-08-31 #4672 @Rupali-Sahu
Hi Lucy,

Thanks for the reminder. It had completely slipped my mind ever since I
started my new job.
I am kinda stuck since this is not a traditional review process. I have not
really used Github before, so I'm confused about where to run the commands
and create a review report. If there's documentation that can guide me,
that would be great.

Warm regards,
Rupali

On Wed, Aug 31, 2022 at 8:43 AM Lucy Whalley ***@***.***>
wrote:

> @StanczakDominik <https://github.com/StanczakDominik> - I can see you
> have started your review - great! - let me know if you have any queries as
> you progress.
>
> @Rupali-Sahu <https://github.com/Rupali-Sahu> - your first review task is
> to generate a checklist - there is more information at the top of this
> thread. Please, if you have any questions, just ask in the thread below.
>
> I find reviews work best when it is a back-and-forth conversations between
> reviewers and authors, so please feel free to give your thoughts and
> suggestions as you are working through the checklist. In many cases, it may
> best to raise these as an issue on the HallThruster repository - if you do,
> please let us know in this thread.
>
