+++ src/simulation/postprocess.jl
@@ -259,8 +259,8 @@ function frame_dict(sol, frame)
         "ui_2" => (ncharge > 1 ? sol[:ui, 2][frame] : filler),
         "ui_3" => (ncharge > 2 ? sol[:ui, 3][frame] : filler),
         "niui_1" => sol[:niui, 1][frame],
-        "niui_2" => (ncharge > 1 ? sol[:niui, 1][frame] : filler),
-        "niui_3" => (ncharge > 2 ? sol[:niui, 1][frame] : filler),
+        "niui_2" => (ncharge > 1 ? sol[:niui, 2][frame] : filler),
+        "niui_3" => (ncharge > 2 ? sol[:niui, 3][frame] : filler),
         "ue" => sol[:ue][frame],
         "V" => sol[:ϕ][frame],
         "E" => sol[:E][frame],
