### 2022-09-05 #4672 @lucydot
Hi Rupali - there is some more documentation here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html.

Github can be a bit daunting when you first use it, however you only need to think about the issue thread (there is a lot more github functionality which isn't used in this process).

You can run the commands in the Github thread itself (rather than via email) - the thread address is here: https://github.com/openjournals/joss-reviews/issues/4672. There are more instructions at the top of that page - please ask me more questions as they arise.
