### 2022-09-11 #4672 @Rupali-Sahu
This is a good initiative and I hope it can be extended to incorporate other plasma dynamics and higher dimensions as well.

I am working remotely on a company laptop which doesn't allow third party installations, so I cannot comment on the functionality or documentation of the code. I hope that is not a problem. I have a few questions/concerns regarding the computational model and the physics being resolved:

1. Line 25: The electron flux causing the problem is actually parallel to the applied electric field. The 'Hall' current in the direction perpendicular to both magnetic and electric fields (i.e. in the azimuthal direction), but the anomalous electron transport is primarily in the axial direction.

2. An explanation is needed on why the lesser-known flux schemes are used instead of the more standard schemes like HLLC, Steger-Warming, Roe's method, etc.

3. I'm not sure of JOSS, but in general it is expected that all the governing equations are presented in the paper itself, instead of referring to the documentation. It would also help to add the details of the three subcases used for comparison. Please ignore this if irrelevant.

4. Details for the fluid model used for comparing the results should be mentioned to some extent. It is unclear at this point if the fluid models are expected to be more or less accurate than the code developed in this study.

5. In general, a code/software presented must be evaluated for its accuracy and limitations. While the comparison results are presented, no discussion is made on the why the differences occur, and in what cases would this code not provide reliable results at all. For instance, removing the anomalous collision would bring the Knudsen number dangerously close to 1, and a fluid description is inaccurate. Also mention under what circumstances does a 1D approximation not hold for a hall thruster.

6. The list of references seems incomplete. Many researchers have worked on fluid/hybrid models of plasma for hall thrusters. It would be helpful to place the current code in context of the existing codes.
