### 2022-09-12 #4672 @lucydot
Hi thanks @Rupali-Sahu for your feedback so far. RE:

> I'm not sure of JOSS, but in general it is expected that all the governing equations are presented in the paper itself, instead of referring to the documentation. It would also help to add the details of the three subcases used for comparison. Please ignore this if irrelevant.

This is not necessary for a JOSS submission.

@StanczakDominik are you able to take a look at this code over the next few weeks?
