### 2022-09-12 #4672 @archermarx
Hi Rupali! Thank you for your feedback. Let me try and address some of the concerns. Note that many of these were not included in the paper due to the word limit. I don't know if you had a chance to look through the documentation, but many of these ought to be covered there.

>     1. Line 25: The electron flux causing the problem is actually parallel to the applied electric field. The 'Hall' current in the direction perpendicular to both magnetic and electric fields (i.e. in the azimuthal direction), but the anomalous electron transport is primarily in the axial direction.
>

Yes, this is a typo. It should say "perpendicular to the applied magnetic field"

>     2. An explanation is needed on why the lesser-known flux schemes are used instead of the more standard schemes like HLLC, Steger-Warming, Roe's method, etc.

I'm not sure these are necessarily lesser-known. Rusanov/Local Lax Friedrichs is a particularly well-known scheme in CFD, particularly within education, and the Global Lax Friedrichs is a straightforward modification of this scheme. I do agree it would be good to discuss why these fluxes were chosen, but due to wordcount limitations,  I think most of this would have to be relegated to the documentation. Adding more fluxes is definitely something we'd like to do.

>     4. Details for the fluid model used for comparing the results should be mentioned to some extent. It is unclear at this point if the fluid models are expected to be more or less accurate than the code developed in this study.
>
That would be good to include. The LANDMARK benchmark code should be somewhat less accurate than our code, but it difficult to compare in detail because no numerical information is provided with the benchmark.

>     5. In general, a code/software presented must be evaluated for its accuracy and limitations. While the comparison results are presented, no discussion is made on the why the differences occur, and in what cases would this code not provide reliable results at all. For instance, removing the anomalous collision would bring the Knudsen number dangerously close to 1, and a fluid description is inaccurate. Also mention under what circumstances does a 1D approximation not hold for a hall thruster.
>
Agree. I think more discussion is warranted, but again I am unsure whether that belongs in the paper or in the documentation of the code.

>     6. The list of references seems incomplete. Many researchers have worked on fluid/hybrid models of plasma for hall thrusters. It would be helpful to place the current code in context of the existing codes.
Yes. I think adding more references would definitely be good. @lucydot do references count toward the word limit?
