### 2022-09-16 c59e889
try fixing plotting to work with loaded landmark data
