+++ src/simulation/postprocess.jl
@@ -204,6 +204,7 @@ function load_landmark_data(case, suffix; ncells = 100)
         ne = ne_itp.(zs),
         ∇ϕ = -E_itp.(zs),
         ϕ = ϕ_itp.(zs),
+        nn_tot = nn,
     )
 
     ionization_reactions = HallThruster.load_reactions(LandmarkIonizationLookup(), [Xenon(0), Xenon(1)]);
