### 2022-09-19 b2d123b
fix LANDMARK plotting
