+++ src/simulation/postprocess.jl
@@ -202,6 +202,9 @@ function load_landmark_data(case, suffix; ncells = 100)
         Tev = 2/3 * ϵ_itp.(zs),
         pe = ϵ_itp.(zs),
         ne = ne_itp.(zs),
+        ni = [ne_itp.(zs)]' |> collect,
+        ui = [ui]' |> collect,
+        niui = [ui]' |> collect,
         ∇ϕ = -E_itp.(zs),
         ϕ = ϕ_itp.(zs),
         nn_tot = nn,
