+++ test/run.jl
@@ -1,4 +1,4 @@
-using Test, HallThruster, Plots, StaticArrays, DiffEqCallbacks, LinearAlgebra, DiffEqBase, LoopVectorization
+using Test, HallThruster, Plots, StaticArrays, DiffEqCallbacks, LinearAlgebra, LoopVectorization
 using OrdinaryDiffEq, PartialFunctions, SpecialFunctions
 
 
@@ -8,7 +8,7 @@ function run_sim(duration = 0.0002; ncells = 50, nsave = 2, dt = 1e-8,
         alg = SSPRK22(stage_limiter! = HallThruster.stage_limiter!, step_limiter! = HallThruster.stage_limiter!),
         flux = HallThruster.rusanov, ionization_model = HallThruster.LandmarkIonizationLookup(), transition = HallThruster.LinearTransition(0.001, 0.0),
         coupled = true, LANDMARK = true,
-        progress_interval = 0, WENO = false, L = 0.05
+        progress_interval = 0, L = 0.05
     )
 
     un = 150.0
@@ -23,7 +23,7 @@ function run_sim(duration = 0.0002; ncells = 50, nsave = 2, dt = 1e-8,
         (0.5, 1.0)
     elseif case == 3
         (0.4, 1.0)
-    elseif case == 4 
+    elseif case == 4
         (0.15, 1.0)
     elseif case == 5
         (0.1, 1.0)
@@ -31,7 +31,7 @@ function run_sim(duration = 0.0002; ncells = 50, nsave = 2, dt = 1e-8,
 
     αw = 1.0
 
-    scheme = HallThruster.HyperbolicScheme(flux, limiter, reconstruct, WENO)
+    scheme = HallThruster.HyperbolicScheme(flux, limiter, reconstruct)
 
     ϵ_anode = 3.0
     ϵ_cathode = 3.0
