### 2022-09-22 7e7806a
Tests pass, compilation time dramatically reduced
Want to fix tutorial and make sure we still agree with LANDMARK
