+++ test/unit_tests/test_conservation_laws.jl
@@ -180,7 +180,7 @@
             U_euler = hcat(euler_state_L, euler_state, euler_state_R)
 
             index = (ρi = [1], ρiui = [2])
-            scheme = HallThruster.HyperbolicScheme(identity, HallThruster.no_limiter, false, false)
+            scheme = HallThruster.HyperbolicScheme(identity, HallThruster.no_limiter, false)
             config = (;scheme, ncharge = 1)
             params = (;index, config)
 
@@ -190,7 +190,7 @@
             @test edge_L[:, 2] == euler_state
             @test edge_R[:, 1] == euler_state
 
-            scheme = HallThruster.HyperbolicScheme(identity, HallThruster.no_limiter, true, false)
+            scheme = HallThruster.HyperbolicScheme(identity, HallThruster.no_limiter, true)
             HallThruster.compute_edge_states!(edge_L, edge_R, U_euler, params)
             @test edge_L[:, 1] == euler_state_L
             @test edge_R[:, end] == euler_state_R
@@ -202,13 +202,13 @@
             limiters = [
                 HallThruster.koren,
                 HallThruster.minmod,
-                HallThruster.osher,
+                HallThruster.osher(1.5),
                 HallThruster.van_albada,
                 HallThruster.van_leer
             ]
 
             for limiter in limiters
-                scheme = HallThruster.HyperbolicScheme(identity, limiter, true, false)
+                scheme = HallThruster.HyperbolicScheme(identity, limiter, true)
                 HallThruster.compute_edge_states!(edge_L, edge_R, U_euler_2, params)
                 @test edge_L[:, 1] == euler_state_L2
                 @test edge_R[:, end] == euler_state_R
