### 2022-09-22 1f9a0a4
remove StaticArrays and SpecialFunctions
"using" time now down to 2 seconds
