+++ test/unit_tests/test_conservation_laws.jl
@@ -44,15 +44,15 @@
         mXe = Xenon.m
 
         continuity_eq = Fluid(Xe_0, ContinuityOnly(; u, T))
-        continuity_state = SA[ρ]
+        continuity_state = (ρ,)
         continuity = (continuity_state, continuity_eq)
 
         isothermal_eq = Fluid(Xe_0, IsothermalEuler(T))
-        isothermal_state = SA[ρ, ρ * u]
+        isothermal_state = (ρ, ρ * u,)
         isothermal = (isothermal_state, isothermal_eq)
 
         euler_eq = Fluid(Xe_0, EulerEquations())
-        euler_state = SA[ρ, ρ * u, ρ * ϵ]
+        euler_state = (ρ, ρ * u, ρ * ϵ,)
         euler = (euler_state, euler_eq)
 
         laws = [continuity, isothermal, euler]
@@ -131,13 +131,13 @@
 
         @testset "Flux computation" begin
             p = ρ * R * T
-            f_euler = SA[ρ * u, ρ * u^2 + p, ρ * u * (ϵ + p / ρ)]
-            @test flux(continuity...) == SA[f_euler[1]]
-            @test flux(isothermal...) == SA[f_euler[1], f_euler[2]]
+            f_euler = (ρ * u, ρ * u^2 + p, ρ * u * (ϵ + p / ρ),)
+            @test flux(continuity...) == (f_euler[1],)
+            @test flux(isothermal...) == (f_euler[1], f_euler[2],)
             @test flux(euler...) == f_euler
 
-            isothermal_state_3 = SA[isothermal_state_2[1], -3 * isothermal_state[2]]
-            euler_state_3 = SA[euler_state_2[1], -2 * euler_state_2[2], euler_state_2[3]]
+            isothermal_state_3 = (isothermal_state_2[1], -3 * isothermal_state[2],)
+            euler_state_3 = (euler_state_2[1], -2 * euler_state_2[2], euler_state_2[3])
 
             # upwind flux
             @test upwind(continuity_state, continuity_state_2, continuity_eq) == flux(continuity...)
@@ -166,8 +166,8 @@
             Te = 6.0
             pe = HallThruster.e * ne * Te
             @test flux(continuity..., pe) ==  flux(continuity...)
-            @test flux(isothermal..., pe) == flux(isothermal...) + SA[0.0, pe]
-            @test flux(euler..., pe) == flux(euler...) + SA[0.0, pe, 0.0]
+            @test flux(isothermal..., pe) == flux(isothermal...) + (0.0, pe,)
+            @test flux(euler..., pe) == flux(euler...) + (0.0, pe, 0.0,)
         end
 
         @testset "Reconstruction" begin
