+++ src/numerics/flux.jl
@@ -4,33 +4,76 @@ Base.@kwdef struct HyperbolicScheme{F}
     reconstruct::Bool = true
 end
 
-function flux(U::SVector{1, T}, fluid, pe = 0.0) where T
+function flux(U::NTuple{1, T}, fluid, pe = 0.0) where T
     ρ = U[1]
     u = velocity(U, fluid)
-    return SA[ρ * u]
+    return (ρ * u,)
 end
 
-function flux(U::SVector{2, T}, fluid, pe = 0.0) where T
+function flux(U::NTuple{2, T}, fluid, pe = 0.0) where T
     ρ, ρu = U
     u = velocity(U, fluid)
     p = pressure(U, fluid)
 
-    return SA[ρu, ρu * u + p + pe]
+    return (ρu, ρu * u + p + pe)
 end
 
-function flux(U::SVector{3, T}, fluid, pe = 0.0) where T
+function flux(U::NTuple{3, T}, fluid, pe = 0.0) where T
     ρ, ρu, ρE = U
     u = U[2] / U[1]
     p = pressure(U, fluid)
     ρH = ρE + p
-    return SA[ρu, ρu * u + p + pe, ρH * u]
+    return (ρu, ρu * u + p + pe, ρH * u)
+end
+
+macro NTuple(ex)
+    if !isa(ex, Expr)
+        error("Bad input for @NTuple")
+    end
+    head = ex.head
+    if head === :comprehension
+        if length(ex.args) != 1 || !isa(ex.args[1], Expr) || ex.args[1].head != :generator
+            error("Expected generator in comprehension, e.g. [f(i) for i = 1:3]")
+        end
+        ex = ex.args[1]
+        if length(ex.args) != 2
+            error("Use a one-dimensional comprehension for @$SV")
+        end
+        rng = eval(ex.args[2].args[2])
+        exprs = (:(f($j)) for j in rng)
+        return quote
+            let
+                f($(esc(ex.args[2].args[1]))) = $(esc(ex.args[1]))
+                NTuple{$(length(rng))}($tuple($(exprs...)))
+            end
+        end
+    elseif head === :typed_comprehension
+        if length(ex.args) != 2 || !isa(ex.args[2], Expr) || ex.args[2].head != :generator
+            error("Expected generator in typed comprehension, e.g. Float64[f(i) for i = 1:3]")
+        end
+        T = esc(ex.args[1])
+        ex = ex.args[2]
+        if length(ex.args) != 2
+            error("Use a one-dimensional comprehension for @$SV")
+        end
+        rng = eval(ex.args[2].args[2])
+        exprs = (:(f($j)) for j in rng)
+        return quote
+            let
+                f($(esc(ex.args[2].args[1]))) = $(esc(ex.args[1]))
+                NTuple{$(length(rng)),$T}($tuple($(exprs...)))
+            end
+        end
+    else
+        error("Expected comprehension")
+    end
 end
 
 # use fun metaprogramming create specialized flux versions for each type of fluid
 for NUM_CONSERVATIVE in 1:3
     eval(quote
 
-    function upwind(UL::SVector{$NUM_CONSERVATIVE, T}, UR::SVector{$NUM_CONSERVATIVE, T}, fluid::Fluid, coupled = false, TeL = 0.0, TeR = 0.0, neL = 1.0, neR = 1.0, args...) where T
[…]
