+++ src/physics/thermodynamics.jl
@@ -7,17 +7,17 @@
 @inline number_density(U, f::Fluid) = density(U, f) / m(f)
 @inline density(U, f::Fluid) = U[1]
 
-@inline velocity(U::SVector{1, T}, f::Fluid) where T = f.conservation_laws.u
-@inline velocity(U::SVector{2, T}, f::Fluid) where T = U[2] / U[1]
-@inline velocity(U::SVector{3, T}, f::Fluid) where T = U[2] / U[1]
+@inline velocity(U::NTuple{1, T}, f::Fluid) where T = f.conservation_laws.u
+@inline velocity(U::NTuple{2, T}, f::Fluid) where T = U[2] / U[1]
+@inline velocity(U::NTuple{3, T}, f::Fluid) where T = U[2] / U[1]
 
-@inline temperature(U::SVector{1, T}, f::Fluid) where T = f.conservation_laws.T
-@inline temperature(U::SVector{2, T}, f::Fluid) where T = f.conservation_laws.T
-@inline temperature(U::SVector{3, T}, f::Fluid) where T = (γ(f) - 1) * (U[3] - 0.5 * U[2]^2 / U[1]) / U[1] / R(f)
+@inline temperature(U::NTuple{1, T}, f::Fluid) where T = f.conservation_laws.T
+@inline temperature(U::NTuple{2, T}, f::Fluid) where T = f.conservation_laws.T
+@inline temperature(U::NTuple{3, T}, f::Fluid) where T = (γ(f) - 1) * (U[3] - 0.5 * U[2]^2 / U[1]) / U[1] / R(f)
 
-@inline pressure(U::SVector{1, T}, f::Fluid) where T = U[1] * R(f) * f.conservation_laws.T
-@inline pressure(U::SVector{2, T}, f::Fluid) where T = U[1] * R(f) * f.conservation_laws.T
-@inline pressure(U::SVector{3, T}, f::Fluid) where T = (γ(f) - 1) * (U[3] - 0.5 * U[2]^2 / U[1])
+@inline pressure(U::NTuple{1, T}, f::Fluid) where T = U[1] * R(f) * f.conservation_laws.T
+@inline pressure(U::NTuple{2, T}, f::Fluid) where T = U[1] * R(f) * f.conservation_laws.T
+@inline pressure(U::NTuple{3, T}, f::Fluid) where T = (γ(f) - 1) * (U[3] - 0.5 * U[2]^2 / U[1])
 
 function stagnation_energy(U, f::Fluid)
     if f.conservation_laws.type == _EulerEquations

[…]
