### 2022-09-23 7d1b821
fix test bugs and remove some test deps
