### 2022-09-23 cd18333
fix bugs resulting from removal of some test deps
