### 2022-09-28 161286b
update efficiency computations and add them to JSON output
