### 2022-09-29 1cea796
fix unending compilation and add minor performance improvements
