### 2022-09-30 acfbb76
Minor performance improvements, code still compiles properly
