### 2022-09-30 c7f7b21
more performance improvements, need to make tests pass
