### 2022-09-30 aba8c93
fix wall loss tests
