+++ test/unit_tests/test_walls.jl
@@ -12,7 +12,11 @@
     Tev = 4.0
     ne = 1e18
     cache = (;
-        ne = [ne, ne, ne, ne], Tev = [Tev, Tev, Tev, Tev], Z_eff = [1.0, 1.0, 1.0, 1.0]
+        ne = [ne, ne, ne, ne],
+        Tev = [Tev, Tev, Tev, Tev],
+        Z_eff = [1.0, 1.0, 1.0, 1.0],
+        ni = [ne ne ne ne],
+        γ_SEE = [0.0, 0.0, 0.0, 0.0]
     )
     transition_function = HallThruster.StepFunction()
     config = (;
@@ -27,7 +31,36 @@
 
     grid = HallThruster.generate_grid(HallThruster.geometry_SPT_100, 2, (0, 2 * L_ch))
 
-    params = (;cache, config, index, z_cell = grid.cell_centers, z_edge = grid.edges, L_ch = L_ch, A_ch = A_ch)
+    z_cell = grid.cell_centers
+    z_edge = grid.edges
+
+    # Fill up cell lengths and magnetic field vectors
+    Δz_cell = zeros(length(z_cell))
+    Δz_edge = zeros(length(z_edge))
+    for (i, z) in enumerate(grid.cell_centers)
+        if firstindex(z_cell) < i < lastindex(z_cell)
+            Δz_cell[i] = z_edge[HallThruster.right_edge(i)] - z_edge[HallThruster.left_edge(i)]
+        elseif i == firstindex(z_cell)
+            Δz_cell[i] = z_edge[begin+1] - z_edge[begin]
+        elseif i == lastindex(z_cell)
+            Δz_cell[i] = z_edge[end] - z_edge[end-1]
+        end
+    end
+
+    for i in eachindex(z_edge)
+        Δz_edge[i] = z_cell[i+1] - z_cell[i]
+    end
+
+
+    mi_kr = HallThruster.Krypton.m
+    γmax = 1 - 8.3 * sqrt(me/mi)
+    γmax_kr = 1 - 8.3 * sqrt(me/mi_kr)
+
+    params = (;
+        cache, config, index, z_cell = grid.cell_centers,
+        z_edge = grid.edges, L_ch = L_ch, A_ch = A_ch,
+        Δz_edge, Δz_cell, γ_SEE_max = γmax
+    )
 
     ρi = ne * mi
     nϵ = 3/2 * ne * Tev
@@ -48,11 +81,6 @@
     @test HallThruster.sheath_potential(Tev, γ2, mi) == Tev*log(0.5*(1-γ2)*sqrt(2*mi/π/me))
     #-1.02 * Tev
 
-    mi_kr = HallThruster.Krypton.m
-
-    γmax = 1 - 8.3 * sqrt(me/mi)
-    γmax_kr = 1 - 8.3 * sqrt(me/mi_kr)
-
     ideal_dielectric = HallThruster.IdealDielectric
     @test HallThruster.SEE_yield(ideal_dielectric, 100.0, γmax) == 0.0
     BN = HallThruster.BoronNitride
@@ -64,6 +92,7 @@
     @test HallThruster.SEE_yield(BN, 9000.0, γmax_kr) ≈ γmax_kr
 
     γ = HallThruster.SEE_yield(BN, Tev, γmax)
+
     Vs = HallThruster.sheath_potential(Tev, γ, mi)
 
     α = 1/4
@@ -80,8 +109,10 @@
     @test HallThruster.freq_electron_wall(landmark_losses, U, params, 3) == 0.0e7
 
     νiw = α * 2 / Δr * sqrt(HallThruster.e * Tev / mi)
-    γ = HallThruster.SEE_yield(BN, Tev, mi)
[…]
