### 2022-09-30 b936b56
further minor performance improvements
