+++ src/HallThruster.jl
@@ -7,7 +7,6 @@ using SparseArrays: Tridiagonal
 using PartialFunctions
 using QuadGK: quadgk
 using DelimitedFiles: readdlm, writedlm
-#using SpecialFunctions: erf
 using FunctionWrappers: FunctionWrapper
 using Unitful: @u_str, uconvert, ustrip, Quantity
 import SnoopPrecompile
