+++ src/simulation/simulation.jl
@@ -54,13 +54,18 @@ function allocate_arrays(grid, fluids) #rewrite allocate arrays as function of s
     Id  = [0.0]
     Vs = [0.0]
 
+
+    # timestepping caches
     k = copy(U)
     u1 = copy(U)
+
+    # other caches
+    cell_cache_1 = zeros(ncells)
 
     cache = (;
                 Aϵ, bϵ, B, νan, νc, μ, ϕ, ∇ϕ, ne, Tev, pe, ue, ∇pe,
                 νen, νei, νew, νiw, νe, F, UL, UR, Z_eff, λ_global, νiz, νex, K, Id, ji,
-                ni, ui, Vs, niui, nn, nn_tot, k, u1, γ_SEE
+                ni, ui, Vs, niui, nn, nn_tot, k, u1, γ_SEE, cell_cache_1
             )
 
     return U, cache
