### 2022-10-19 b2892a2
update electron wall loss model
