+++ src/HallThruster.jl
@@ -58,7 +58,7 @@ include("simulation/potential.jl")
 include("simulation/heavy_species.jl")
 include("simulation/electronenergy.jl")
 include("simulation/sourceterms.jl")
-include("simulation/update_values.jl")
+include("simulation/update_electrons.jl")
 include("simulation/configuration.jl")
 include("simulation/restart.jl")
 include("simulation/ssprk_step.jl")
