### 2022-10-20 73abd3a
Fix restart loading
