+++ src/simulation/restart.jl
@@ -108,6 +108,14 @@ function load_restart(grid, fluids, config, sol::Solution)
             for Z in 1:config.ncharge
                 @. @views cache[field][Z, :] = U[2 * Z + 1, :] / U[2 * Z, :]
             end
+        elseif field == :niui
+            for Z in 1:config.ncharge
+                @. @views cache[field][Z, :] = U[2 * Z + 1, :] / config.propellant.m
+            end
+        elseif field == :nn
+            for i in 1:1+config.solve_background_neutrals
+                @. @views cache[field][i, :] = U[i, :] / config.propellant.m
+            end
         else
             itp = LinearInterpolation(z_cell, sv[field])
             cache[field] .= itp.(grid.cell_centers)
