### 2022-11-02 235db90
fix tests for altered wall loss model
