+++ test/unit_tests/test_walls.jl
@@ -16,7 +16,8 @@
         Tev = [Tev, Tev, Tev, Tev],
         Z_eff = [1.0, 1.0, 1.0, 1.0],
         ni = [ne ne ne ne],
-        γ_SEE = [0.0, 0.0, 0.0, 0.0]
+        γ_SEE = [0.0, 0.0, 0.0, 0.0],
+        νew = [0.0, 0.0, 0.0, 0.0],
     )
     transition_function = HallThruster.StepFunction()
     config = (;
@@ -81,10 +82,8 @@
     @test HallThruster.sheath_potential(Tev, γ2, mi) == Tev*log(0.5*(1-γ2)*sqrt(2*mi/π/me))
     #-1.02 * Tev
 
-    ideal_dielectric = HallThruster.IdealDielectric
-    @test HallThruster.SEE_yield(ideal_dielectric, 100.0, γmax) == 0.0
     BN = HallThruster.BoronNitride
-    @test HallThruster.SEE_yield(BN, 10.0, γmax) ≈ BN.Γ * 10.0^BN.b * BN.a
+    @test HallThruster.SEE_yield(BN, 10.0, γmax) ≈ BN.σ₀ + 1.5 * 10.0/BN.ϵ_star * (1 - BN.σ₀)
     # Check that SEE properly obeys space charge limit at high electron temps
     @test HallThruster.SEE_yield(BN, 9000.0, γmax) ≈ γmax
 
@@ -97,7 +96,6 @@
 
     α = 1/4
     sheath_model = HallThruster.WallSheath(BN, α)
-
     Δr = HallThruster.geometry_SPT_100.outer_radius - HallThruster.geometry_SPT_100.inner_radius
     Δz = params.z_edge[2] - params.z_edge[1]
     V_cell = A_ch * Δz
@@ -108,11 +106,15 @@
     @test HallThruster.freq_electron_wall(landmark_losses, U, params,  2) == 1e7
     @test HallThruster.freq_electron_wall(landmark_losses, U, params, 3) == 0.0e7
 
-    νiw = α * 2 / Δr * sqrt(HallThruster.e * Tev / mi)
     γ = HallThruster.SEE_yield(BN, Tev, γmax)
+    νiw = α * sqrt(HallThruster.e * Tev / mi) / Δr * γ
     νew = νiw / (1 - γ)
 
     params.cache.γ_SEE .= γ
+    params.cache.νew[1] = νew
+    params.cache.νew[2] = νew
+    params.cache.νew[3] = 0.0
+    params.cache.νew[4] = 0.0
 
     Iiw = HallThruster.wall_ion_current(sheath_model, U, params, 2, 1)
     Iew = HallThruster.wall_electron_current(sheath_model, U, params, 2)
@@ -123,8 +125,8 @@
     @test HallThruster.freq_electron_wall(sheath_model, U, params, 2) ≈ νew
     @test HallThruster.freq_electron_wall(sheath_model, U, params, 3) ≈ 0.0
 
-    @test HallThruster.wall_power_loss(sheath_model, U, params, 2) ≈ νew * (2 * Tev[1] + (1 - γ) * Vs)
-    @test HallThruster.wall_power_loss(sheath_model, U, params, 3) ≈ 0.0
+    @test HallThruster.wall_power_loss(sheath_model, U, params, 2) ≈ νew * (2 * (1 - 0.5 * BN.σ₀) * Tev + (1 - γ) * Vs)/γ
+    @test HallThruster.wall_power_loss(sheath_model, U, params, 4) ≈ 0.0
 end
 
[…]
