+++ paper/paper.md
@@ -35,16 +35,17 @@ The Hall effect thruster is an  annular plasma device characterized by an electr
 
 ![Hall thruster operation\label{fig:hallthruster}](hallthruster.png)
 
-The heavier, unmagnetized ions are accelerated by this electric field downstream at high speed while the electrons remain trapped in a strong, $E\times B$  azimuthal "Hall" current.  In an ideal, collisionless environment, these trapped electrons are not able to cross the confining magnetic field.  In practice, however, there is a high degree of electron flux across the magnetic field lines, perpendicular to the applied electric field. The process, which is a critical driver for the overall dynamics of the thruster, is not well understood. `HallThruster.jl` was designed as a tool to more rapidly assess different  models for this electron transport through parameter estimation, uncertainty quantification, and surrogate optimization.   It allows users to alter many parts of the physics model for electron transport (and related processes) without modifying the source code of the package, and the one-dimensional domain only requires a few seconds to a few minutes to run a simulation, depending on the grid resolution. With that said, a central premise of Hallthruster.jl is that the dynamics in the Hall thruster are suitably one-dimensional that the 1D code is usually sufficient for an accurate approximation of a 2D or 3D code when run on the scale of the entire thruster.  But, we note that with appropriate calibration, errors may be corrected by training a surrogate model correction for higher dimensional codes.  Indeed, this is one of the anticipated use cases. 
+The heavier, unmagnetized ions are accelerated by this electric field downstream at high speed while the electrons remain trapped in a strong, $E\times B$  azimuthal "Hall" current.  In an ideal, collisionless environment, these trapped electrons are not able to cross the confining magnetic field.  In practice, however, there is a high degree of electron flux across the magnetic field lines, perpendicular to the applied magnetic field. The process, which is a critical driver for the overall dynamics of the thruster, is not well understood. `HallThruster.jl` was designed as a tool to more rapidly assess different  models for this electron transport through parameter estimation, uncertainty quantification, and surrogate optimization.   It allows users to alter many parts of the physics model for electron transport (and related processes) without modifying the source code of the package, and the one-dimensional domain only requires a few seconds to a few minutes to run a simulation, depending on the grid resolution. With that said, a central premise of Hallthruster.jl is that the dynamics in the Hall thruster are suitably one-dimensional that the 1D code is usually sufficient for an accurate approximation of a 2D or 3D code when run on the scale of the entire thruster.  But, we note that with appropriate calibration, errors may be corrected by training a surrogate model correction for higher dimensional codes.  Indeed, this is one of the anticipated use cases.
 
-While one-dimensional codes are commonly used in Hall thruster research [@haraquasineutralfluid] [@sahu_ffm] [@mikellides_1D],  `HallThruster.jl` is unique in that, in addition to its primary goal of informing closure models, it is written from the ground up with a strong emphasis on code verification, collaboration, and extension.  As an open source code, our goal is that it will allow the community to reproduce computational results obtained and to compare and contrast with parallel  studies.
+While one-dimensional codes are commonly used in Hall thruster research [@boeuf1998] [@ahedo2001] [@haraquasineutralfluid] [@sahu_ffm] [@mikellides_1D],  `HallThruster.jl` is unique in that, in addition to its primary goal of informing closure models, it is written from the ground up with a strong emphasis on code verification, collaboration, and extension.  As an open source code, our goal is that it will allow the community to reproduce computational results obtained and to compare and contrast with parallel studies.
 
 # Physics model
 
[…]
