+++ paper/paper.bib
@@ -8,13 +8,13 @@ pages = {113505},
 year = {2020},
 doi = {10.1063/5.0021474},
 
-URL = { 
+URL = {
         https://doi.org/10.1063/5.0021474
-    
+
 },
-eprint = { 
+eprint = {
         https://doi.org/10.1063/5.0021474
-    
+
 }
 
 }
@@ -31,7 +31,6 @@ URL = {https://arc.aiaa.org/doi/abs/10.2514/6.2001-3505},
 eprint = {https://arc.aiaa.org/doi/pdf/10.2514/6.2001-3505}
 }
 
-
 @article{Hara_2019,
 	doi = {10.1088/1361-6595/ab0f70},
 	url = {https://doi.org/10.1088/1361-6595/ab0f70},
@@ -45,6 +44,37 @@ eprint = {https://arc.aiaa.org/doi/pdf/10.2514/6.2001-3505}
 	title = {An overview of discharge plasma modeling for Hall effect thrusters},
 	journal = {Plasma Sources Science and Technology},
 	abstract = {The current status and challenges of the discharge plasma modeling capabilities in Hall effect thrusters (HETs) are reviewed. Partially magnetized plasmas in HETs present unique physics in that collisions with neutral atoms (a feature of low-temperature plasmas) and instabilities and oscillations (a feature of high-temperature plasmas) coexist. A discharge plasma model is constructed by choosing the numerical methods, e.g. fluid or kinetic, for individual plasma constituents, including neutral atoms, ions, and electrons. The governing equations, boundary conditions, and collision models of a xenon discharge plasma in HETs are discussed. Various models are used to illustrate device-scale phenomena, including the breathing mode and azimuthally rotating spokes, as well as small-scale physics, such as the electron cyclotron drift instability and current-carrying instability that may occur in a hollow cathode plume.}
+}
+
+@article{boeuf1998,
+abstract = {Stationary plasma thrusters are ion thrusters whose properties make them especially suitable for satellite station keeping or orbit transfer. In these thrusters, a magnetic field transverse to the electron flow towards the anode increases the electron collision frequency and makes possible the generation of a plasma at relatively low gas flow and gas.density. The decrease of the plasma conductivity due to the magnetic field induces a large electric field in the plasma which accelerates the quasicollisionless ions whose trajectories are not significantly affected by the magnetic field. The purpose of this article is to clarify, using results from a simple model, the electrical properties of these thrusters and the low frequency oscillation regime which has been observed experimentally. The model is based on the assumption of quasineutrality of the plasma column and on a 1D transient hybrid treatment of electron and ion transport in the device. Electrons are considered as a fluid and ions are described with a collisionless kinetic equation. This model provides reasonable estimates of the plasma properties and is able to give a clear picture of the low frequency oscillations, qualitatively close to the experimental observations. {\textcopyright} 1998 American Institute of Physics.},
+author = {Boeuf, J. P. and Garrigues, L.},
+doi = {10.1063/1.368529},
+file = {:C\:/Users/thoma/AppData/Local/Mendeley Ltd./Mendeley Desktop/Downloaded/Boeuf, Garrigues - 1998 - Low frequency oscillations in a stationary plasma thruster(2).pdf:pdf},
+issn = {00218979},
+journal = {Journal of Applied Physics},
+month = {oct},
+number = {7},
+pages = {3541--3554},
+publisher = {American Institute of Physics Inc.},
+title = {{Low frequency oscillations in a stationary plasma thruster}},
+volume = {84},
+year = {1998}
+}
+
+@article{ahedo2001,
+abstract = {},
[…]
