### 2022-11-08 #4672 @lucydot
@StanczakDominik I think mid-November is fine, especially given that we are struggling to find alternative reviewers - thank you for the update. Although all checkboxes on the reviewer list are important, particular emphasis should be given to `Functionality` as @Rupali-Sahu has not been able to verify this.

@Rupali-Sahu I can see there are some items still ticked off on your list that do not require installation of the software (under `Documentation` and `Software paper`). Please do continue to work through these items and tick-off if/when met. @archermarx  has updated the paper since you last updated the review thread (see link to paper above).
