### 2022-11-26 #4672 @Rupali-Sahu
Hi Lucy!

I am happy with the author's response on my comments. I would recommend this work for publication in JOSS as it would provide an easily accessible tool to test many hypotheses for anomalous electron transport.
