### 2022-12-05 #4672 @Rupali-Sahu
Hi Lucy! I updated my checklist. I could also download and run the software on my new personal laptop so I have ticked off a few boxes in the documentation and functionality section as well. In the documentation,
1. Few equations, such as that for electron heat conduction and two-zone Bohm collision frequency are not printing in math mode, possibly a missing bracket or something like that.
2. In the tutorial, it is unclear how an anomalous electron transport model is selected.

I could not find a statement of need and community guidelines in the documentation. If the authors could add that, or point me to where its located, I can tick off the corresponding boxes. Other than that, the documentation looks quite detailed and informative :)
