+++ src/simulation/initialization.jl
@@ -78,9 +78,6 @@ function initialize!(U, params, ::DefaultInitialization)
 
         U[index.nϵ, i] = number_density_function(z) * energy_function(z)
     end
-
-    # Initialize the anomalous collision frequency
-    initialize_anom!(params.cache.νan, params.config.anom_model, U, params)
-
+    
     return U
 end
