### 2022-12-08 9789b34
anom model interface updated, still need to run tests and update docs
