+++ src/collisions/anomalous.jl
@@ -1,99 +1,37 @@
 """
     AnomalousTransportModel
-The abstract supertype of all types of anomalous transport model. Currently, there are two subtypes: `FixedAnomModel` and `ZeroEquationModel`, which
-correspond, respectively, to models which are initialized and then never updated, and models which depend algebraically on local plasma properties.
+The abstract supertype of all types of anomalous transport models.
+Subtype this to define your own model.
 """
 abstract type AnomalousTransportModel end
 
-"""
-    FixedAnomModel <: AnomalousTransportModel
-An anomalous collision frequency model which is fixed at the start of a simulation and then never updated again.
-This is the most common type of model, and includes the ubiquitous multi-zone bohm-like collision frequency models.
-
-Users defining their own `FixedAnomModel` must implement the `initialize_anom!(νan, model, U, params)` method for their type.
-"""
-abstract type FixedAnomModel <: AnomalousTransportModel end
-
-"""
-    ZeroEquationModel <: AnomalousTransportModel
-An anomalous collision frequency which depends algebraically on the local plasma properties and is updated at each iteration.
-
-Users defining their own `ZeroEquationModel` must implement the `evaluate_anom(model, U, params)` method for their type.
-"""
-abstract type ZeroEquationModel <: AnomalousTransportModel end
-
-"""
-    initialize_anom!(νan, model::FixedAnomModel, U, params)
-Initialize the anomalous collision frequency profile for models in which the anomalous mobility does not change with time. Must be
-implemented by the user for specific subtypes of FixedAnomModel. Should update argument νan in-place with the desired value of anomalous
-collision frequency and then return the mutated array.
-"""
-function initialize_anom!(νan, model::FixedAnomModel, U, params)
-    throw(ArgumentError(
-        "Function `initialize_anom!(νan, model, U, params)` is not implemented for model of type $(typeof(model)). This function must be implemented by the user for `FixedAnomModel` objects."
-    ))
-end
-
-"""
-    evaluate_anom(model::FixedAnomModel, U, params, i)
-By default, this does nothing for FixedAnomModel objects, as the anomalous collision frequency is not updated at each iteration.
-"""
-function evaluate_anom(model::FixedAnomModel, U, params, i)
-    return params.cache.νan[i]
-end
-
-
-"""
-    initialize_anom!(νan, model::ZeroEquationModel, U, params)
-Default anomalous collision frequency initialization for `ZeroEquationModel`s. Just calls evalute_anom at each grid location.
-"""
-function initialize_anom!(νan, model::ZeroEquationModel, U, params)
-    for i in eachindex(params.z_cell)
-        νan[i] = evaluate_anom(model, U, params, i)
-    end
-    return νan
-end
-
-"""
-    evaluate_anom(model::ZeroEquationModel, U, params, i)
-Return the anomalous collision frequency at index `i` for a provided ZeroEquationModel. Must be implemented by the user.
-"""
-function evaluate_anom(model::ZeroEquationModel, U, params, i)
-    throw(ArgumentError(
-        "Function `evaluate_anom(model, U, params, i)` is not implemented for model of type $(typeof(model)). This function must be implemented by the user for `ZeroEquationModel` objects."
-    ))
-end
-
 #=============================================================================
- Begin definition of built-in models of type FixedAnomModel
+ Begin definition of built-in models
 ==============================================================================#
 
 """
-    NoAnom <: FixedAnomModel
[…]
