### 2022-12-22 #4672 @archermarx
I will also be on holiday until around then. Apologies for the delay, busy
time of year for me.

Thomas Marks

Ph.D. Candidate
Department of Aerospace Engineering
Plasmadynamics and Electric Propulsion Laboratory
University of Michigan, Ann Arbor
Email: ***@***.***
Phone: 480-532-1189
Office: FXB 1216 G-5

On Wed, Dec 21, 2022 at 4:34 AM Lucy Whalley ***@***.***>
wrote:

> Hi @archermarx <https://github.com/archermarx>, any upates or timescales
> for completion?
>
> @StanczakDominik <https://github.com/StanczakDominik> a reminder for your
> review, or timescale for when you will be able to review this?
>
> Also a note here that I will be on holiday from today until the 4th of
> January 🎄 though I may check in between the mince pies, we'll see!
>
