+++ test/runtests.jl
@@ -7,71 +7,88 @@ using DelimitedFiles
 using LinearAlgebra
 using Unitful
 using SparseArrays
+using Alert
 
-doctest(HallThruster)
-
-include("unit_tests/test_gas.jl")
-include("unit_tests/test_conservation_laws.jl")
-include("unit_tests/test_limiters.jl")
-include("unit_tests/test_reactions.jl")
-include("unit_tests/test_misc.jl")
-include("unit_tests/test_electrons.jl")
-include("unit_tests/test_boundary_conditions.jl")
-include("unit_tests/test_walls.jl")
-include("unit_tests/test_initialization.jl")
-include("unit_tests/test_restarts.jl")
-
-@testset "Order verification (electron energy)" begin
-    include("order_verification/ovs_funcs.jl")
-    include("order_verification/ovs_energy.jl")
-
-    vfunc = x -> OVS_Energy.verify_energy(x)
-    refinements = refines(6, 20, 2)
-
-    # Test spatial order of accuracy of implicit solver and crank-nicholson on L1, L2, and L∞ norms
-    slopes_nϵ, norms_nϵ = test_refinements(vfunc, refinements, [1, 2, Inf])
-
-    tol = 0.15
-    order = 1.0
-
-    # Check that electron energy is solved to at least first order
-    for slope in slopes_nϵ
-        @show slope
-        @test abs(slope - order) < tol || slope ≥ order
+function try_alert(msg)
+    try
+        alert(msg)
+    catch
     end
+    return nothing
 end
 
-@testset "Order verification (neutrals and ions)" begin
-    include("order_verification/ovs_funcs.jl")
-    include("order_verification/ovs_ions.jl")
-    refinements = refines(5, 40, 2)
+try
+    doctest(HallThruster)
 
-    limiter = HallThruster.van_leer
-    flux_names = ("HLLE", "Rusanov", "Global Lax-Friedrichs")
-    fluxes = (HallThruster.HLLE, HallThruster.rusanov, HallThruster.global_lax_friedrichs,)
+    include("unit_tests/test_gas.jl")
+    include("unit_tests/test_conservation_laws.jl")
+    include("unit_tests/test_limiters.jl")
+    include("unit_tests/test_reactions.jl")
+    include("unit_tests/test_misc.jl")
+    include("unit_tests/test_electrons.jl")
+    include("unit_tests/test_boundary_conditions.jl")
+    include("unit_tests/test_walls.jl")
+    include("unit_tests/test_initialization.jl")
+    include("unit_tests/test_restarts.jl")
 
-    for (flux, flux_name) in zip(fluxes, flux_names)
-        for reconstruct in (false, true)
+    @testset "Order verification (electron energy)" begin
+        include("order_verification/ovs_funcs.jl")
+        include("order_verification/ovs_energy.jl")
 
-            if reconstruct && flux_name == "HLLE"
-                continue
-            end
+        vfunc = x -> OVS_Energy.verify_energy(x)
+        refinements = refines(6, 20, 2)
 
-            scheme = HallThruster.HyperbolicScheme(flux, limiter, reconstruct)
+        # Test spatial order of accuracy of implicit solver and crank-nicholson on L1, L2, and L∞ norms
+        slopes_nϵ, norms_nϵ = test_refinements(vfunc, refinements, [1, 2, Inf])
 
-            slopes, norms = test_refinements(ncells -> OVS_Ions.solve_ions(ncells, scheme, false), refinements, [1, 2, Inf])
+        tol = 0.15
+        order = 1.0
 
-            # Check that gradient reconstruction gets us to at least ~1.75-order accuracy
-            theoretical_order = 1 + reconstruct
-            for slope in slopes
-                if !reconstruct
[…]
