+++ test/Project.toml
@@ -1,4 +1,5 @@
 [deps]
+Alert = "28312eec-4d86-447d-83ad-bc2b262de792"
 DelimitedFiles = "8bb1440f-4735-579b-a4ab-409b98df4dab"
 Documenter = "e30172f5-a6a5-5a46-863b-614d45cd2de4"
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
