### 2023-01-16 #4672 @archermarx
If i might suggest alternate/additional reviewer(s), Maryam Reza (m.reza20@imperial.ac.uk) and Farbod Faraji (f.faraji20@imperial.ac.uk) both recently published work using a Hall thruster code written in Julia. I think either would be well-qualified to review the code.
