### 2023-01-16 4bbb804
fix allocations problem in conductivity lookup and improve JSON handling
