+++ src/simulation/simulation.jl
@@ -231,71 +231,143 @@ function run_simulation(config::Config;
     return sol
 end
 
-function run_simulation(json_path::String)
-    (;design, simulation, parameters) = JSON3.read(read(json_path, String))
+function run_simulation(json_path::String; single_section = false, is_path = true, nonstandard_keys = false)
 
-    geometry = Geometry1D(;
-        channel_length = design.channel_length,
-        outer_radius = design.outer_radius,
-        inner_radius = design.inner_radius,
-    )
+    if is_path
+        json_content = JSON3.read(read(json_path, String))
+    else
+        json_content = JSON3.read(json_path)
+    end
 
-    bfield_data = readdlm(design.magnetic_field_file, ',')
+
+    if single_section
+
+        if nonstandard_keys
+            (;
+                # Design
+                thruster_name,
+                channel_length, inner_radius, outer_radius, 
+                magnetic_field_file, magnetically_shielded,
+                propellant_material, wall_material,
+                anode_potential, cathode_potential,
+                anode_mass_flow_rate,
+                # Simulation
+                anom_model, cathode_location_m, max_charge,
+                num_cells, dt_s, duration_s, num_save,
+                flux_function, limiter, reconstruct,
+                ion_wall_losses, electron_ion_collisions, solve_background_neutrals,
+                # Parameters
+                anom_coeff_1, anom_coeff_2, sheath_loss_coefficient,
+                ion_temp_K, neutral_temp_K, neutral_velocity_m_s,
+                cathode_electron_temp_eV, inner_outer_transition_length_m,
+                background_pressure_Torr, background_temperature_K,
+            ) = json_content
+
+            anom_model_coeffs = [anom_coeff_1, anom_coeff_2]
+            propellant = propellant_material
+        else
+            (;
+                # Design
+                thruster_name,
+                channel_length, inner_radius, outer_radius, 
+                magnetic_field_file, magnetically_shielded,
+                propellant, wall_material,
+                anode_potential, cathode_potential,
+                anode_mass_flow_rate,
+                # Simulation
+                anom_model, cathode_location_m, max_charge,
+                num_cells, dt_s, duration_s, num_save,
+                flux_function, limiter, reconstruct,
+                ion_wall_losses, electron_ion_collisions, solve_background_neutrals,
+                # Parameters
+                anom_model_coeffs, sheath_loss_coefficient,
+                ion_temp_K, neutral_temp_K, neutral_velocity_m_s,
+                cathode_electron_temp_eV, inner_outer_transition_length_m,
+                background_pressure_Torr, background_temperature_K,
+            ) = json_content
+        end
+    else
+        (;design, simulation, parameters) = json_content
+        (;
+            thruster_name,
+            channel_length, inner_radius, outer_radius, 
+            magnetic_field_file, magnetically_shielded,
+            propellant, wall_material,
+            anode_potential, cathode_potential,
+            anode_mass_flow_rate,
+        ) = design
+
+        (;
+            anom_model, cathode_location_m, max_charge,
+            num_cells, dt_s, duration_s, num_save,
+            flux_function, limiter, reconstruct,
+            ion_wall_losses, electron_ion_collisions, solve_background_neutrals,
+        ) = simulation
+
+        (; 
+            anom_model_coeffs, sheath_loss_coefficient,
+            ion_temp_K, neutral_temp_K, neutral_velocity_m_s,
[…]
