### 2023-01-16 37311dd
fix bohm condition at left boundary
