+++ src/HallThruster.jl
@@ -1,5 +1,6 @@
 module HallThruster
 
+using SpecialFunctions
 using LinearAlgebra
 using DocStringExtensions
 
@@ -9,6 +10,8 @@ using QuadGK: quadgk
 using DelimitedFiles: readdlm, writedlm
 using FunctionWrappers: FunctionWrapper
 using Unitful: @u_str, uconvert, ustrip, Quantity
+
+
 import SnoopPrecompile
 
 using JSON3
