### 2023-01-18 8852464
new ion loss model and lots of minor changes
