+++ src/utilities/utility_functions.jl
@@ -114,36 +114,6 @@ end
 
 @inline ion_current_density(U, p, i) = sum(Z * e * U[p.index.ρiui[Z], i] for Z in 1:p.config.ncharge) / p.config.propellant.m
 
-function discharge_current(U::Matrix, params)
-    (;A_ch, cache, Δz_edge, ϕ_L, ϕ_R) = params
-    (;∇pe, μ, ne, ji, Vs) = cache
-
-    ncells = size(U, 2)
-
-    int1 = 0.0
-    int2 = 0.0
-
-    @inbounds for i in 1:ncells-1
-        Δz = Δz_edge[i]
-
-        int1_1 = (ji[i] / e / μ[i] + ∇pe[i]) / ne[i]
-        int1_2 = (ji[i+1] / e / μ[i+1] + ∇pe[i+1]) / ne[i+1]
-
-        int1 += 0.5 * Δz * (int1_1 + int1_2)
-
-        int2_1 = inv(e * ne[i] * μ[i])
-        int2_2 = inv(e * ne[i+1] * μ[i+1])
-
-        int2 += 0.5 * Δz * (int2_1 + int2_2)
-    end
-
-    ΔV = ϕ_L + Vs[] - ϕ_R
-
-    I = A_ch * (ΔV + int1) / int2
-
-    return I
-end
-
 function cumtrapz(x, y, y0 = zero(typeof(y[1] * x[1])))
     int = zeros(typeof(y[1] * x[1]), length(x))
     cumtrapz!(int, x, y, y0)
