+++ src/simulation/ssprk_step.jl
@@ -34,7 +34,8 @@ function solve(prob::ODEProblem; saveat, dt)
     fields_to_save = (
         :μ, :Tev, :ϕ, :∇ϕ, :ne, :pe, :ue, :∇pe, :νan, :νc, :νen,
         :νei, :νew, :νiz, :νex, :νe, :Id, :ni, :ui, :ji, :niui, :nn, :nn_tot,
-        :anom_multiplier, :ohmic_heating, :wall_losses, :inelastic_losses, :Vs
+        :anom_multiplier, :ohmic_heating, :wall_losses, :inelastic_losses, :Vs,
+        :channel_area, :inner_radius, :outer_radius, :dA_dz, :tanδ
     )
 
     first_saveval = NamedTuple{fields_to_save}(params.cache)
@@ -46,11 +47,15 @@ function solve(prob::ODEProblem; saveat, dt)
     while t < tspan[2]
         i += 1
         t += dt
+
         # Update heavy species
         ssprk22_step!(u, f, params, t)
 
         # Update electron quantities
         update_electrons!(u, params, t)
+
+        # Update plume geometry
+        update_plume_geometry!(u, params)
 
         # Check for NaNs and terminate if necessary
         nandetected = false
