+++ src/HallThruster.jl
@@ -61,6 +61,7 @@ include("simulation/potential.jl")
 include("simulation/heavy_species.jl")
 include("simulation/electronenergy.jl")
 include("simulation/sourceterms.jl")
+include("simulation/plume.jl")
 include("simulation/update_electrons.jl")
 include("simulation/configuration.jl")
 include("simulation/restart.jl")
