### 2023-01-24 #4672 @archermarx
Thank you Lucy!

Thomas Marks

Ph.D. Candidate
Department of Aerospace Engineering
Plasmadynamics and Electric Propulsion Laboratory
University of Michigan, Ann Arbor
Email: ***@***.***
Phone: 480-532-1189
Office: FXB 1216 G-5

On Tue, Jan 24, 2023 at 6:42 AM Lucy Whalley ***@***.***>
wrote:

> A note that I have also contacted Farbod to ask if available to review.
>
