### 2023-01-30 f87dda9
add config option for solving plume
