+++ src/simulation/configuration.jl
@@ -45,6 +45,7 @@ struct Config{A<:AnomalousTransportModel, W<:WallLossModel, IZ<:IonizationModel,
     background_neutral_temperature::Float64
     anode_boundary_condition::Symbol
     anom_smoothing_iters::Int
+    solve_plume::Bool
 end
 
 function Config(;
@@ -87,6 +88,7 @@ function Config(;
         background_neutral_temperature      = 100.0u"K",
         anode_boundary_condition            = :sheath,
         anom_smoothing_iters                = 0,
+        solve_plume = false
     ) where {IC, S_N, S_IC, S_IM, S_ϕ, S_E}
 
     # check that number of ion source terms matches number of charges for both
@@ -148,7 +150,8 @@ function Config(;
         background_pressure,
         background_neutral_temperature,
         anode_boundary_condition,
-        anom_smoothing_iters
+        anom_smoothing_iters,
+        solve_plume
     )
 end
 
