### 2023-01-31 bb1e788
fix tests and bump version
