### 2023-02-02 5e3e41d
Remove debug print statement in loading json file
