### 2023-02-17 #4672 @archermarx
Here are some other people i found in the JOSS list of reviewers that may be suitable

benjaminbolling
rangsimanketkaew
pkairys
amorenobr
p-ortmann
TomGoffrey
joglekara
CFGrote
sigvaldm
rogeriojorge
michaeltouati
akaptano
OsAmaro
hasantahir
