### 2023-06-18 #4672 @lucydot
Hi @archermarx I've just started working through the checklist. Quick one - I couldn't find guidelines for "Report issues or problems with the software" or "Seek support". This could be fixed with a couple of sentences on the "Home" page of your documentation, for example.
