### 2023-06-18 #4672 @lucydot
Paper:

- should `one-dimensional fluid code` be `one-dimensional fluid dynamics code`?
- you should be able to place the list of references on page 2 into a single bracket: `[@boeuf1998; @ahedo2001; @haraquasineutralfluid; @sahu_ffm; @mikellides_1D]`
- `.(Thomas, 1949)` --> ` (Thomas, 1949)`(additional period removed)
- The `Landmark 1D fluid test case` weblink is broken (in reference list). Please be sure to use correct [citation syntax](https://pandoc.org/MANUAL.html#extension-citations).
- I think Gowda et al has now been peer reviewed+published (currently referencing the arxiv paper)
