### 2023-06-18 #4672 @lucydot
Tutorials:

- I had to add `show=true` when using `plot!` within a user-defined function. For example, in the tutorial function `plot_current`: `plot!(p, t_ms, ion_current, label = "Ion current (cathode)") --> `plot!(p, t_ms, ion_current, label = "Ion current (cathode)",show=true)` (would also need updating at other places).

It may be that your system is setup to output plot as default. I am using a fresh install of Julia and IJulia. May catch others out so may be worth making a note of behaviour if don't want to add additional keyword argument to every `plot` call (which does feel clunky way to handle).

- There are some warnings towards the end of the tutorial. Is this acceptable? May be worth making a note if so, or fix if not.
