### 2023-06-19 #4672 @archermarx
> Hi @archermarx I've just started working through the checklist.
>
>     * I couldn't find guidelines for "Report issues or problems with the software" or "Seek support". This could be fixed with a couple of sentences on the "Home" page of your documentation, for example.
>
>     * Link to tutorial from the README is broken
>
>     * PR with very small update to install instruction here: [Update installation instructions UM-PEPL/HallThruster.jl#83](https://github.com/UM-PEPL/HallThruster.jl/pull/83) (tested on Julia 1.9)

I've accepted your PR and made the relevant changes.

> Tutorials:
>
>     * I had to add `show=true` when using `plot!` within a user-defined function. For example, in the tutorial function `plot_current`: `plot!(p, t_ms, ion_current, label = "Ion current (cathode)") --> `plot!(p, t_ms, ion_current, label = "Ion current (cathode)",show=true)` (would also need updating at other places).
>
>
> It may be that your system is setup to output plot as default. I am using a fresh install of Julia and IJulia. May catch others out so may be worth making a note of behaviour if don't want to add additional keyword argument to every `plot` call (which does feel clunky way to handle).
>
>     * There are some warnings towards the end of the tutorial. Is this acceptable? May be worth making a note if so, or fix if not.

Thanks for this. I added a note about the warnings to the tutorial and changed the plot parts of the tutorial to explicitly call `display`.
