### 2023-06-21 #4672 @lucydot
Hi @archermarx - my previous PR missed an @ sign in the citation list, I think this caused it not to format correctly. Just put in a PR that fixes. Can you merge? https://github.com/UM-PEPL/HallThruster.jl/pull/84
