### 2023-06-28 #4672 @kyleniemeyer
Congratulations @archermarx on your article's publication in JOSS! Please consider [signing up as a reviewer](https://reviewers.joss.theoj.org/join) to help us in the future, if you haven't already.
